feat: convert now constant to TZDate + useDateRanges hook
